+++ Sources/KeyboardKit/Callouts/View+InputCallout.swift
@@ -15,7 +15,7 @@ public extension View {
      able to present an input callout.
      */
     func inputCallout(
-        context: InputCalloutContext = .shared,
+        context: InputCalloutContext,
         style: InputCalloutStyle = .standard) -> some View {
         ZStack {
             self