+++ Sources/KeyboardKit/System/SystemKeyboardSpaceButton.swift
@@ -31,11 +31,12 @@ public struct SystemKeyboardSpaceButton: View {
     private let actionHandler: KeyboardActionHandler
     private var action: KeyboardAction { .space }
     
-    @EnvironmentObject var context: ObservableKeyboardContext
+    @EnvironmentObject private var context: ObservableKeyboardContext
+    @EnvironmentObject private var inputCalloutContext: InputCalloutContext
     
     public var body: some View {
         SystemKeyboardSpaceButtonContent(localeText: localeText, spaceText: spaceText)
             .standardButtonStyle(for: action, context: context)
-            .keyboardGestures(for: action, actionHandler: actionHandler)
+            .keyboardGestures(for: action, actionHandler: actionHandler, inputCalloutContext: inputCalloutContext)
     }
 }